correct sensor name
