+++ README.md
@@ -29,7 +29,7 @@ There is:
    * [`esp32-c3-dkc02-bsc`](./common/lib/esp32-c3-dkc02-bsc) - Board-Support for the ESP32-C3-DKC02
    * [`get-uuid`](./common/lib/get-uuid) - provides a compile-time generated UUID
    * [`mqtt-messages`](./common/lib/mqtt-messages) - MQTT helper functions
-   * [`imc42670p`](./common/lib/imc42670p) - basic sensor driver
+   * [`icm42670p`](./common/lib/icm42670p) - basic sensor driver
 * Some extra bits:
    * [`mqtt-python-client`](./extra/mqtt-python-client) A Python MQTT client, for testing
