+++ advanced/i2c-driver/exercise/src/icm42670p.rs
@@ -2,75 +2,73 @@
 
 use embedded_hal::blocking::i2c;
 
-/// IMC42670P device driver.
+/// ICM42670P device driver, represented by a struct with 2 fields.
 /// Datasheet: https://3cfeqx1hf82y3xcoull08ihx-wpengine.netdna-ssl.com/wp-content/uploads/2021/07/DS-000451-ICM-42670-P-v1.0.pdf
 #[derive(Debug)]
-pub struct IMC42670P<I2C> {
+pub struct ICM42670P<I2C> {
     /// The concrete I²C device implementation.
-    i2c: I2C,
-
+    //  TODO! field 1 
     /// Device address
-    address: DeviceAddr,
+    //  TODO! field 2 
 }
 
 // See Table 3.3.2 in Documentation
 /// Contains the possible variants of the devices addesses as binary numbers.
 #[derive(Debug, Clone, Copy, PartialEq)]
-pub enum DeviceAddr {
+// TODO! Implement the public enum with the two adress variants here
 
-    /// 0x68
-    AD0 = 0b110_1000,
-    /// 0x69
-    AD1 = 0b110_1001,
-}
 
-impl<I2C, E>IMC42670P<I2C>
+
+
+
+// impl block with methods
+impl<I2C, E>ICM42670P<I2C>
 where
+    // this defines which error messages will be used
     I2C: i2c::WriteRead<Error = E> + i2c::Write<Error = E>,
 {
     /// Creates a new instance of the sensor, taking ownership of the i2c peripheral.
     pub fn new(i2c: I2C, address: DeviceAddr) -> Result<Self, E> {
-
-        let imc42670p = IMC42670P { i2c, address };
-
-        Ok(imc42670p)
+        // instantiates the ICM42670P struct 
+        // returns the struct as Ok value
+        // TODO!
     }
 
     /// Returns the device's ID `0x67 
     //(if it doesn't, something is amiss)
     // Public method that can be accessed from outside this file.
     pub fn read_device_id_register(&mut self) -> Result<u16, E> {
-        self.read_register(Register::WhoAmI)
+        // reads the Device ID register
+        // TODO!
     }
 
     /// Writes into a register
     // This method is not public as it is only needed inside this file.
     #[allow(unused)]
     fn write_register(&mut self, register: Register, value: u8) -> Result<(), E> {
-        let byte = value as u8;
-        self.i2c
-            .write(self.address as u8, &[register.address(), byte])
+        // value that will be written as u8
+        // i2c write 
+        // TODO!
     }
 
     /// Reads a register using a `write_read` method.
     // This method is not public as it is only needed inside this file.
     fn read_register(&mut self, register: Register) -> Result<u16, E> {
-        let mut data = [0; 2];
-        self.i2c
-            .write_read(self.address as u8, &[register.address()], &mut data)?;
-        Ok(u16::from_le_bytes(data))
-    }
+        // buffer for values
+        // i2c write_read
+        // return u16 from ke bytes
+        // TODO!
 }
 
 // See Table 14.1 in documentation
 /// This enum represents the device's registers
 #[derive(Clone, Copy)]
 pub enum Register {
-    WhoAmI = 0x75,
+    WhoAmI = // // TODO! Fill in the register
 }
 
 impl Register {
     fn address(&self) -> u8 {
-        *self as u8
+        // TODO! Returns Register as u8
     }
 }
