+++ advanced/i2c-driver/exercise/src/main.rs
@@ -5,9 +5,9 @@ use esp_idf_hal::{
     prelude::*,
 };
 use esp_idf_sys::*;
-use i2c_driver_exercise::imc42670p::{IMC42670P, DeviceAddr};
+use i2c_driver_exercise::icm42670p::{ICM42670P, DeviceAddr};
 
-// Dont change this file. Work in the imc42670p.rs and modify it so main.rs runs.
+// Dont change this file. Work in the icm42670p.rs and modify it so main.rs runs.
 
 fn main() -> anyhow::Result<()> {
     link_patches();
@@ -23,7 +23,7 @@ fn main() -> anyhow::Result<()> {
         <MasterConfig as Default>::default().baudrate(400.kHz().into()),
     )?;
 
-    let mut sensor = IMC42670P::new(i2c, DeviceAddr::AD0)?;
+    let mut sensor = ICM42670P::new(i2c, DeviceAddr::AD0)?;
     println!("Sensor init");
     let device_id = sensor.read_device_id_register()?;
