+++ advanced/i2c-driver/solution/src/icm42670p.rs
@@ -2,73 +2,75 @@
 
 use embedded_hal::blocking::i2c;
 
-/// IMC42670P device driver, represented by a struct with 2 fields.
+/// ICM42670P device driver.
 /// Datasheet: https://3cfeqx1hf82y3xcoull08ihx-wpengine.netdna-ssl.com/wp-content/uploads/2021/07/DS-000451-ICM-42670-P-v1.0.pdf
 #[derive(Debug)]
-pub struct IMC42670P<I2C> {
+pub struct ICM42670P<I2C> {
     /// The concrete I²C device implementation.
-    //  TODO! field 1 
+    i2c: I2C,
+
     /// Device address
-    //  TODO! field 2 
+    address: DeviceAddr,
 }
 
 // See Table 3.3.2 in Documentation
 /// Contains the possible variants of the devices addesses as binary numbers.
 #[derive(Debug, Clone, Copy, PartialEq)]
-// TODO! Implement the public enum with the two adress variants here
+pub enum DeviceAddr {
 
+    /// 0x68
+    AD0 = 0b110_1000,
+    /// 0x69
+    AD1 = 0b110_1001,
+}
 
-
-
-
-// impl block with methods
 impl<I2C, E>ICM42670P<I2C>
 where
-    // this defines which error messages will be used
     I2C: i2c::WriteRead<Error = E> + i2c::Write<Error = E>,
 {
     /// Creates a new instance of the sensor, taking ownership of the i2c peripheral.
     pub fn new(i2c: I2C, address: DeviceAddr) -> Result<Self, E> {
-        // instantiates the ICM42670P struct 
-        // returns the struct as Ok value
-        // TODO!
+
+        let icm42670p = ICM42670P { i2c, address };
+
+        Ok(icm42670p)
     }
 
     /// Returns the device's ID `0x67 
     //(if it doesn't, something is amiss)
     // Public method that can be accessed from outside this file.
     pub fn read_device_id_register(&mut self) -> Result<u16, E> {
-        // reads the Device ID register
-        // TODO!
+        self.read_register(Register::WhoAmI)
     }
 
     /// Writes into a register
     // This method is not public as it is only needed inside this file.
     #[allow(unused)]
     fn write_register(&mut self, register: Register, value: u8) -> Result<(), E> {
-        // value that will be written as u8
-        // i2c write 
-        // TODO!
+        let byte = value as u8;
+        self.i2c
+            .write(self.address as u8, &[register.address(), byte])
     }
 
     /// Reads a register using a `write_read` method.
     // This method is not public as it is only needed inside this file.
     fn read_register(&mut self, register: Register) -> Result<u16, E> {
-        // buffer for values
-        // i2c write_read
-        // return u16 from ke bytes
-        // TODO!
+        let mut data = [0; 2];
+        self.i2c
+            .write_read(self.address as u8, &[register.address()], &mut data)?;
+        Ok(u16::from_le_bytes(data))
+    }
 }
 
 // See Table 14.1 in documentation
 /// This enum represents the device's registers
 #[derive(Clone, Copy)]
 pub enum Register {
-    WhoAmI = // // TODO! Fill in the register
+    WhoAmI = 0x75,
 }
 
 impl Register {
     fn address(&self) -> u8 {
-        // TODO! Returns Register as u8
+        *self as u8
     }
 }
