+++ advanced/i2c-driver/solution/src/lib.rs
@@ -1,4 +1,4 @@
 #![deny(unsafe_code)]
 #![no_std]
 
-pub mod imc42670p;
+pub mod icm42670p;
