+++ advanced/i2c-driver/solution/src/main.rs
@@ -5,7 +5,7 @@ use esp_idf_hal::{
     prelude::*,
 };
 use esp_idf_sys::*;
-use i2c_driver_exercise::imc42670p::{IMC42670P, SlaveAddr};
+use i2c_driver_exercise::icm42670p::{ICM42670P, SlaveAddr};
 
 // Dont change this file. Work in the lib.rs and modify it so main.rs runs.
 
@@ -26,9 +26,9 @@ fn main() -> anyhow::Result<()> {
         <MasterConfig as Default>::default().baudrate(400.kHz().into()),
     )?;
 
-    let mut sensor = IMC42670P::new(i2c, SlaveAddr::AD0)?;
+    let mut sensor = ICM42670P::new(i2c, SlaveAddr::AD0)?;
     // If you are using an ESP32-C3-DevKitC-02, change to:
-    // let mut sensor = IMC42670P::new(i2c, SlaveAddr::AD1)?;
+    // let mut sensor = ICM42670P::new(i2c, SlaveAddr::AD1)?;
     println!("Sensor init");
     let device_id = sensor.read_device_id_register()?;
