+++ book/src/04_3_2_i2c.md
@@ -2,7 +2,7 @@
 
 We're not going to write an entire driver, merely the first step: the `hello world` of driver writing: reading the device ID of the sensor. This version is labelled easy, because we explain the code fragments, and you only have to copy and paste the fragments into into the right place. Use this version if you have very little previous experience with Rust, if these workshops are your first in the embedded domain, or if you found the hard version too hard. 
 
-`i2c-driver/exercise/src/imc42670p.rs` is a gap text of a very basic I²C IMU sensor driver. The task is to complete the file, so that running `main.rs` will log the device ID of the driver. The this gap text driver is based on the version of the same name that lives in common, but provides a little bit more functionality.
+`i2c-driver/exercise/src/icm42670p.rs` is a gap text of a very basic I²C IMU sensor driver. The task is to complete the file, so that running `main.rs` will log the device ID of the driver. The this gap text driver is based on the version of the same name that lives in common, but provides a little bit more functionality.
 
 ## Driver API
 
@@ -29,7 +29,7 @@ impl<I2C, E>ICM42670P<I2C>
 where
     I2C: i2c::WriteRead<Error = E> + i2c::Write<Error = E>,
 {
-    /// Create a new instance of the IMC42670P.
+    /// Create a new instance of the ICM42670P.
     pub fn new(i2c: I2C, address: DeviceAddr) -> Result<Self, E> {
 
         let icm42670p = ICM42670P { i2c, address };
@@ -74,7 +74,7 @@ impl Register {
 We define a read and a write method, based on methods provided by the `embedded-hal` crate. They serve as helpers for more specific methods and as an abstraction that is adapted to a sensor with 8-bit registers. Note how the `read_register()` method is based on a `write_read()` method. The reason for this lies in the characteristics of the I²C protocol: We first need to write a command over the I²C bus to specify which register we want to read from. Helper methods can remain private as they don't need to be accessible from outside this crate. 
 
 ```rust
-impl<I2C, E>IMC42670P<I2C>
+impl<I2C, E>ICM42670P<I2C>
 where
     I2C: i2c::WriteRead<Error = E> + i2c::Write<Error = E>,
 {    
@@ -112,5 +112,5 @@ where
 let GYRO_DATA_X: i16 = ((GYRO_DATA_X1 as i16) << 8) | GYRO_DATA_X0 as i16;
 ```
 (maybe from `from_be_bytes[GYRO_data_X1, GYRO_DATA_X2] is better)
-- if you need hints and inspiration, you can check the imc42670p in the common/lib
+- if you need hints and inspiration, you can check the icm42670p in the common/lib
