+++ book/src/04_3_3_i2c.md
@@ -2,7 +2,7 @@
 
 We're not going to write an entire driver, merely the first step: the `hello world` of driver writing: Reading the device ID of the sensor. This version is labelled hard, because you have to come up with the content of the methods and research information in the embedded-hal and data-sheets yourself. 
 
-`i2c-driver/exercise/src/imc42670p.rs` is a gap text of a very basic I²C IMU sensor driver. The task is to complete the file, so that running `main.rs` will log the device ID of the driver. The this gap text driver is based on the version of the same name that lives in common, but provides a little bit more functionality.
+`i2c-driver/exercise/src/icm42670p.rs` is a gap text of a very basic I²C IMU sensor driver. The task is to complete the file, so that running `main.rs` will log the device ID of the driver. The this gap text driver is based on the version of the same name that lives in common, but provides a little bit more functionality.
 
 ## Driver API
 
@@ -36,7 +36,7 @@ impl<I2C, E>ICM42670P<I2C>
 where
     I2C: i2c::WriteRead<Error = E> + i2c::Write<Error = E>,
 {
-    /// Create a new instance of the IMC42670P.
+    /// Create a new instance of the ICM42670P.
     pub fn new(i2c: I2C, address: DeviceAddr) -> Result<Self, E> {
 
         let icm42670p = ICM42670P { i2c, address };
@@ -126,7 +126,7 @@ The reason for this lies in the characteristics of the I²C protocol: We first n
 <Details>
     <Summary>Example Solution</Summary>
 ```rust
-impl<I2C, E>IMC42670P<I2C>
+impl<I2C, E>ICM42670P<I2C>
 where
     I2C: i2c::WriteRead<Error = E> + i2c::Write<Error = E>,
 {    
@@ -153,7 +153,7 @@ where
 <Details>
     <Summary>Example Solution</Summary>
 ```rust
-impl<I2C, E>IMC42670P<I2C>
+impl<I2C, E>ICM42670P<I2C>
 where
     I2C: i2c::WriteRead<Error = E> + i2c::Write<Error = E>,
 {    
@@ -182,5 +182,5 @@ pub fn read_device_id_register(&mut self) -> Result<u16, E> {
 let GYRO_DATA_X: i16 = ((GYRO_DATA_X1 as i16) << 8) | GYRO_DATA_X0 as i16;
 ```
 (maybe from `from_be_bytes[GYRO_data_X1, GYRO_DATA_X2] is better)
-- if you need hints and inspiration, you can check the imc42670p in the common/lib
+- if you need hints and inspiration, you can check the icm42670p in the common/lib
