+++ common/lib/icm42670p/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "imc42670p"
+name = "icm42670p"
 version = "0.1.0"
 authors = ["Tanks Transfeld <tanks.transfeld@ferrous-systems.com>"]
 edition = "2021"
